+++ include/vse/mat4.hpp
@@ -5,11 +5,8 @@ private:
     float values [16];
 public:
     mat4();
-    mat4(float x);
     mat4(float arr[16]);
     ~mat4() = default;
 
     const float *const get_values() const { return values; }
 };
-
-mat4 mult(const mat4& lhs, const mat4& rhs);
